DEV-2130 Remove post step in prd stage of CI/CD
This step fetches all the changes between two git tags. But this step fails
if there is no previous git tag or if git is not installed in the default
Docker container.

It should update the JIRA tickets linked to these changes with the PRD build
info but that doesn't happen (correctly) so just remove this step.