DEV-2130 Fix post step in prd stage of CI/CD (#5)
This step fetches all the changes between two git tags. But this step fails
if there is no previous git tag or if git is not installed in the default
Docker container. It should update the JIRA tickets linked to these changes with
the PRD build info but that doesn't happen (correctly) at the moment. Only the
last ticket will be updated.

For now, just remove the diff between the two tags so that it doesn't crash anymore.
The last ticket will still be updated with the PRD build info.